fixed code and all tests
